pyproject.toml
[tool.poetry]
name = "cafe backend"
version = "0.1.0"
description = "Simple cafe backend."
authors = ["Antony Nyagah <tony.m.nyagah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.16"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.8.1"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
